Add some integration tests
Integration tests are run with $ yarn run integration-test

The integration tests check that the package can be installed and used.
(i.e. the matcher function exists and works as expected). In particular
it checks that the behaviour is correct when invalid XML is compared,
as it's hard to reliably test that in the normal unit tests.

The module is packed into a tgz which is installed in a fresh
environment (unconnected to the project repo), which is used to run the
integration tests.